+++ packages/openapi/src/requests/go.ts
@@ -1,34 +1,39 @@
 'use client';
-import { ident, SampleGenerator } from '@/requests/_shared';
-import { resolveRequestData } from '@/utils/url';
+import { ident, type SampleGenerator } from '@/requests/_shared';
 
 export const generator: SampleGenerator = (url, data, { mediaAdapters }) => {
   const imports = ['fmt', 'net/http', 'io/ioutil'];
   const headers = new Map<string, string>();
   const variables = new Map<string, string>();
-  variables.set('url', JSON.stringify(resolveRequestData(url, data)));
+  variables.set('url', JSON.stringify(url));
 
   for (const header in data.header) {
-    headers.set(header, JSON.stringify(data.header[header]));
+    headers.set(header, JSON.stringify(data.header[header].value));
   }
 
-  const cookies = Object.keys(data.cookie);
-  if (cookies.length > 0)
+  const cookies = Object.entries(data.cookie);
+  if (cookies.length > 0) {
     headers.set(
       'Cookie',
-      JSON.stringify(cookies.map((p) => `${p}=${data.cookie[p]}`).join('; ')),
+      JSON.stringify(
+        cookies.map(([k, param]) => `${k}=${param.value}`).join('; '),
+      ),
     );
+  }
 
   let body: string | undefined;
 
   if (data.body && data.bodyMediaType && data.bodyMediaType in mediaAdapters) {
     headers.set('Content-Type', `"${data.bodyMediaType}"`);
-    body = mediaAdapters[data.bodyMediaType].generateExample(data, {
-      lang: 'go',
-      addImport(from) {
-        imports.push(from);
+    body = mediaAdapters[data.bodyMediaType].generateExample(
+      data as { body: unknown },
+      {
+        lang: 'go',
+        addImport(from) {
+          imports.push(from);
+        },
       },
-    });
+    );
   }
 
   return `package main